Keepalive int64->*int

+++ internal/configs/version2/config.go
@@ -13,7 +13,7 @@ type Upstream struct {
 	Name      string
 	Servers   []UpstreamServer
 	LBMethod  string
-	Keepalive int64
+	Keepalive int
 }
 
 // UpstreamServer defines an upstream server.
